+++ src/com/kaylerrenslow/armaDialogCreator/arma/stringtable/KVPair.java
@@ -1,9 +1,0 @@
-package com.kaylerrenslow.armaDialogCreator.arma.stringtable;
-
-/**
- Created by Kayler on 05/23/2016.
- */
-public interface KVPair<K,V> {
-	K getKey();
-	V getValue();
-}