+++ src/com/kaylerrenslow/armaDialogCreator/arma/stringtable/Language.java
@@ -1,0 +1,13 @@
+package com.kaylerrenslow.armaDialogCreator.arma.stringtable;
+
+import org.jetbrains.annotations.NotNull;
+
+/**
+ @author Kayler
+ @see StringTableValue
+ @since 12/12/2016 */
+public interface Language {
+	/** Get name of language */
+	@NotNull
+	String getName();
+}