+++ src/com/kaylerrenslow/armaDialogCreator/arma/stringtable/StringTableParser.java
@@ -1,0 +1,21 @@
+package com.kaylerrenslow.armaDialogCreator.arma.stringtable;
+
+import org.jetbrains.annotations.NotNull;
+
+import java.io.IOException;
+
+/**
+ Used to construct a new {@link StringTable} instance
+
+ @author Kayler
+ @since 12/12/2016 */
+public interface StringTableParser {
+	/**
+	 Get a new {@link StringTable} instance for the given file
+
+	 @return the new instance
+	 @throws IOException when the table couldn't be parsed
+	 */
+	@NotNull
+	StringTable createStringTableInstance() throws IOException;
+}