+++ src/com/kaylerrenslow/armaDialogCreator/arma/stringtable/StringTableValue.java
@@ -1,0 +1,18 @@
+package com.kaylerrenslow.armaDialogCreator.arma.stringtable;
+
+import org.jetbrains.annotations.NotNull;
+
+import java.util.Map;
+
+/**
+ @author Kayler
+ @since 12/12/2016 */
+public interface StringTableValue {
+
+	@NotNull
+	StringTableKey getKey();
+
+	@NotNull
+	Map<Language, String> getValues();
+
+}